requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
SQLAlchemy==2.0.0
asyncpg==0.27.0
alembic==1.11.2
pydantic==1.10.9
python-dotenv==1.0.0
aiokafka==0.8.0
pytest==8.3.3
pytest-asyncio==0.20.3
httpx==0.24.1
aiosqlite==0.19.0
python-multipart==0.0.18
black==24.10.0
